backend/requirements.txt
langchain-core>=0.3.0,<0.4.0
langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0,<0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0


duckduckgo-search>=7.0.0
sentence-transformers
chromadb>=0.4.0
ddgs

hydra-core
beautifulsoup4
fastapi
uvicorn>=0.20.0
python-multipart
pypdf
PyPDF2
pdfplumber
pypinyin

# For SharedLLM (Qwen2.5-7B-Instruct)
torch>=2.3.0
transformers>=4.40.0
accelerate>=0.30.0
python-dotenv
